+++ CHANGELOG.md
@@ -1,10 +1,19 @@
 # Change Log
 
+## 0.4.1
++ Add output to terminal;
+* Change `response { css.model(), xpath.model(), re.model() }` to `response { cssModel(), xpathModel(), reModel() }`;
+* Fix `response.xpath()` getter likes `/@href` will return `Atrr` instead of its value;
+* Fix `response.css()` selector's getter likes `::html()` will remove content.
+
 ## 0.4.0
 * New Design.
     * concept `Source` inspired by [html-scrapper](https://github.com/harish2704/html-scrapper)
     * `response { css(), xpath(), re() }`, `pipeline` inspired by [scrapy](https://github.com/scrapy/scrapy)
     * `response { css.model(), xpath.model(), re.model() }` inspired by [node-scrapy](https://github.com/eeshi/node-scrapy)
+* Set different handler for extracting data from the page's url which match different patterns;
+* Set same handler for saving data in pipeline;
+* Schedule task fetching in specify time, according to `cron`.
 
 ---